refactor(core): 归一 IntegrityManifest 的 canonical payload 字节路径
- api-contracts/src/lib.rs 给 IntegrityManifest 加权威方法 canonical_payload_bytes，
  落入单一 PayloadView（version/generated_at/files，不含 signature），字段顺序固定
- build-tools/src/lib.rs 删除本地 CanonicalManifest/CanonicalManifestFile 两个投影结构，
  canonical_manifest_payload 改为委托 manifest.canonical_payload_bytes()
- security-core/integrity::ManifestPayload.version 由 u64 对齐为 u32，
  保持与 api-contracts 同一类型；serde_json 输出字节不受影响（整数字面量）
- 新增 build-tools/tests/canonical_bytes_consistency.rs 两条回归：
  1. 跨 crate 字节相等：api_contracts::canonical_payload_bytes ==
     security_core::integrity::canonicalize_manifest(&ManifestPayload)
  2. 快照锁定 canonical JSON 具体字节串（字段顺序+类型序列化规则）
- build-tools/Cargo.toml 加 dev-dependency security_core 以做跨 crate 断言

验证：api_contracts / build-tools / security_core / desktop 全部测试通过
- build-tools 2 单测 + 2 集成测试（含新增的 2 条）
- security_core 38 测试
- desktop bin 82 测试

Made-with: Cursor

+++ backend/shared/security-core/src/integrity.rs
@@ -30,10 +30,12 @@ pub struct ManifestFile {
 /// Manifest payload 主体（去掉 `signature` 之外的所有字段）。
 ///
 /// 字段命名与 Worker 签名侧保持一致；新增字段要同时更新 canonical 序列化。
+/// `version` 与 `api_contracts::IntegrityManifest.version` 对齐为 `u32`，
+/// 便于两侧在同一份 manifest 上比对 canonical 字节串完全一致。
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
 pub struct ManifestPayload {
     /// 格式版本号（打包侧当前恒为 1；日后升级再迭代）。
-    pub version: u64,
+    pub version: u32,
     pub generated_at: String,
     pub files: Vec<ManifestFile>,
 }
